Split namer into its own file
